(Upstream) FB3/DocX/ODT: get lang and description metadata
From upstream: Android: added book annotation (description)
to the book editing dialog.
(cherry picked from commit 28627eab29179e86a569373b8c4d27092b00a010)
