+++ crengine/include/fb3fmt.h
@@ -3,8 +3,26 @@
 
 #include "../include/crsetup.h"
 #include "../include/lvtinydom.h"
+#include "../include/lvopc.h"
 
 bool DetectFb3Format( LVStreamRef stream );
 bool ImportFb3Document( LVStreamRef stream, ldomDocument * doc, LVDocViewCallback * progressCallback, CacheLoadingCallback * formatCallback );
+
+class fb3ImportContext
+{
+private:
+    OpcPackage *m_package;
+    OpcPartRef m_bookPart;
+    ldomDocument *m_descDoc;
+public:
+    fb3ImportContext(OpcPackage *package);
+    virtual ~fb3ImportContext();
+
+    lString16 geImageTarget(const lString16 relationId);
+    LVStreamRef openBook();
+    ldomDocument *getDescription();
+public:
+    lString16 m_coverImage;
+};
 
 #endif // FB3FMT_H
